Added checks before rule removal. Finish associated hashfiles and jobs for customers. Customer removal now removes associated hashfile hashes, hashes, hashfiles and hash notifications... this needs to be tested

+++ hashview/customers/routes.py
@@ -1,6 +1,6 @@
 from flask import Blueprint, render_template, redirect, url_for, flash
-from flask_login import login_required
-from hashview.models import Customers
+from flask_login import login_required, current_user
+from hashview.models import Customers, Jobs, Hashfiles, HashfileHashes, Hashes, HashNotifications
 from hashview.customers.forms import CustomersForm
 from hashview import db
 
@@ -14,7 +14,9 @@ customers = Blueprint('customers', __name__)
 @login_required
 def customers_list():
     customers = Customers.query.all()
-    return render_template('customers.html', title='Cusomters', customers=customers)
+    jobs = Jobs.query.all()
+    hashfiles = Hashfiles.query.all()
+    return render_template('customers.html', title='Cusomters', customers=customers, jobs=jobs, hashfiles=hashfiles)
 
 @customers.route("/customers/add", methods=['GET', 'POST'])
 @login_required
@@ -32,9 +34,29 @@ def customers_add():
 @login_required
 def customers_delete(customer_id):
     customer = Customers.query.get_or_404(customer_id)
-    #if post.author != current_user:  #confirm if admin
-    #    abort(403)
-    db.session.delete(customer)
-    db.session.commit()
-    flash('Customer has been deleted!', 'success')
+    if current_user.admin:
+        # Check if jobs are present
+        jobs = Jobs.query.filter_by(customer_id=customer_id).all()
+        if jobs:
+            flash('Unable to delete. Customer has active job', 'danger')
+        else:
+            # remove associated hash files & hashes & Hash Notifications
+            hashfiles = Hashfiles.query.filter_by(customer_id=customer_id)
+            for hashfile in hashfiles:
+                hashfile_hashes = HashfileHashes.query.filter_by(hashfile_id = hashfile.id).all()
+                for hashfile_hash in hashfile_hashes:
+                    hashes = Hashes.query.filter_by(id=hashfile_hash.id, cracked=0).all()
+                    for hash in hashes:
+                        # Check to see if our hashfile is the ONLY hashfile for this customer that has this hash
+                        customer_cnt = HashfileHashes.query.filter_by(hash_id=hash.id).distinct('customer_id')
+                        if customer_cnt < 2:
+                            db.session.delete(hash)
+                            HashNotifications.query.filter_by(hash_id=hashfile_hash.hash_id).delete()
+                    db.session.delete(hashfile_hash)
+                db.session.delete(hashfile)
+        db.session.delete(customer)
+        db.session.commit()
+        flash('Customer has been deleted!', 'success')
+    else:
+        flash('Permission Denied', 'danger')
     return redirect(url_for('customers.customers_list'))
