+++ hashview/rules/routes.py
@@ -46,8 +46,13 @@ def rules_add():
 def rules_delete(rule_id):
     rule = Rules.query.get(rule_id)
     if current_user.admin or rule.owner_id == current_user.id:
-        db.session.delete(rule)
-        db.session.commit()
+        # Check if part of a task
+        tasks = Tasks.query.filter_by(rule_id=rule.id)
+        if tasks:
+            flash('Rules is currently used in a task and can not be delete.', 'danger')
+        else:
+            db.session.delete(rule)
+            db.session.commit()
         flash('Rule file has been deleted!', 'success')
     else:
         flash('Unauthorized action!', 'danger')
